refactor(lib): extract parseFilters, formatRelative, vote-reducer
Pull three pure helpers out of their component files so they can be
unit-tested without rendering or without awaiting server components:

- lib/filters.ts: BUDGETS, AREAS, HomeSearchParams, and parseFilters
  (previously inline in app/page.tsx). parseFilters now takes a plain
  object, and page.tsx resolves the searchParams Promise itself.
- lib/format.ts: formatRelative(iso, now=Date.now()) consolidating two
  copies that lived in city-detail-reviews and recent-reviews. The
  recent-reviews copy used to round everything >5 weeks as "개월 전";
  it now falls through to years for parity with the detail view, which
  is a strict superset.
- lib/vote-reducer.ts: VoteState, VoteValue, and voteReducer extracted
  from the inline useOptimistic callback in city-card.tsx.

No behavior change beyond the recent-reviews "over a year" rendering.
