+++ app/src/lib/format.ts
@@ -1,0 +1,13 @@
+const DAY_MS = 24 * 60 * 60 * 1000;
+
+export function formatRelative(iso: string, now: number = Date.now()): string {
+  const diffMs = now - new Date(iso).getTime();
+  if (diffMs < DAY_MS) return "오늘";
+  const days = Math.floor(diffMs / DAY_MS);
+  if (days < 7) return `${days}일 전`;
+  const weeks = Math.floor(days / 7);
+  if (weeks < 5) return `${weeks}주 전`;
+  const months = Math.floor(days / 30);
+  if (months < 12) return `${months}개월 전`;
+  return `${Math.floor(days / 365)}년 전`;
+}
